Update -htaccess to fix 1and1 problems
I am using 1&1 for hosting my websites for +6 years. Anyway in the ADK forum you criticised 1and1 :-1: 

In the ADK Forum someone already had a problem.
So the fix for the internal error 500 is to add:
RewriteBase / to the -htaccess in /public.. This fixes every error :)
Working for all versions for BFACP.
Maybe you can add this as a comment to the file?