Fixed #163
If there's just one message, you don't get an array back but just an string. So if $presetMessages is set, but it's not an array, put the string in the first array item. Also the is no 'setting_value', so you would get an 'Trying to get property of non-object' error.